+++ packages/HitTriage/src/app/hit-triage-views/info-view.ts
@@ -16,6 +16,7 @@ import {HitBaseView} from '../base-view';
 export class InfoView extends HitBaseView<HitTriageTemplate, HitTriageApp> {
   public readmePath = _package.webRoot + 'README_HT.md';
   private deletedCampaigns: string[] = [];
+  private dataSourceFunctionsMap: {[key: string]: DG.Func | DG.DataQuery} = {};
   constructor(app: HitTriageApp) {
     super(app);
     this.name = 'Hit Triage';
@@ -30,34 +31,44 @@ export class InfoView extends HitBaseView<HitTriageTemplate, HitTriageApp> {
   }
 
   async init(presetTemplate?: HitTriageTemplate): Promise<void> {
-    const continueCampaignsHeader = ui.h1(i18n.continueCampaigns);
-    const createNewCampaignHeader = ui.h1(i18n.createNewCampaignHeader, {style: {marginLeft: '10px'}});
-    const description = '- Configure your own workflow using the template editor.\n'+
+    ui.setUpdateIndicator(this.root, true);
+    try {
+      const continueCampaignsHeader = ui.h1(i18n.continueCampaigns);
+      const createNewCampaignHeader = ui.h1(i18n.createNewCampaignHeader, {style: {marginLeft: '10px'}});
+      const description = '- Configure your own workflow using the template editor.\n'+
     '- Calculate different molecular properties.\n'+
     '- Filter molecules using different criteria.\n'+
     '- Submit processed dataframe to the function of your choice.\n'+
     '- Save campaigns and continue any time from where you left off.\n ';
-    const appDescription = ui.div([
-      ui.h1('Process, analyse and filter molecules for your needs using Hit Triage:'),
-      ui.div(ui.markdown(description), {style: {color: 'var(--grey-5)'}, classes: 'mb-small'}),
-      ui.link('Read more',
-        'https://github.com/datagrok-ai/public/tree/master/packages/HitTriage'),
-    ]);
+      const appDescription = ui.div([
+        ui.h1('Process, analyse and filter molecules for your needs using Hit Triage:'),
+        ui.div(ui.markdown(description), {style: {color: 'var(--grey-5)'}, classes: 'mb-small'}),
+        ui.link('Read more',
+          'https://github.com/datagrok-ai/public/tree/master/packages/HitTriage'),
+      ]);
 
-    const campaignAccordionDiv = ui.div();
-    const templatesDiv = ui.div([], {classes: 'ui-form'});
+      const campaignAccordionDiv = ui.div();
+      const templatesDiv = ui.div([], {classes: 'ui-form'});
 
-    const campaignsTable = await this.getCampaignsTable();
-    $(this.root).empty();
-    this.root.appendChild(ui.div([
-      appDescription,
-      continueCampaignsHeader,
-      campaignsTable,
-      createNewCampaignHeader,
-      templatesDiv,
-      campaignAccordionDiv,
-    ]));
-    this.startNewCampaign(campaignAccordionDiv, templatesDiv, presetTemplate).then(() => this.app.resetBaseUrl());
+      const campaignsTable = await this.getCampaignsTable();
+
+      await this.startNewCampaign(campaignAccordionDiv, templatesDiv, presetTemplate);
+      $(this.root).empty();
+      this.root.appendChild(ui.div([
+        appDescription,
+        continueCampaignsHeader,
+        campaignsTable,
+        createNewCampaignHeader,
+        templatesDiv,
+        campaignAccordionDiv,
+      ]));
+      this.app.resetBaseUrl();
+    } catch (e) {
+      ui.setUpdateIndicator(this.root, false);
+      throw e;
+    } finally {
+      ui.setUpdateIndicator(this.root, false);
+    }
   }
 
   private async startNewCampaign(
@@ -72,7 +83,7 @@ export class InfoView extends HitBaseView<HitTriageTemplate, HitTriageApp> {
       const templateName = templatesInput.value;
       const template: HitTriageTemplate = presetTemplate && presetTemplate.name === templateName ? presetTemplate :
         JSON.parse(await _package.files.readAsText('Hit Triage/templates/' + templateName + '.json'));
-      const newCampaignAccordeon = await this.getNewCampaignAccordeon(template);
+      const newCampaignAccordeon = await this.getNewCampaignAccordeon(template, containerDiv);
       $(containerDiv).empty();
       containerDiv.className = 'ui-form';
       containerDiv.appendChild(newCampaignAccordeon);
@@ -82,8 +93,10 @@ export class InfoView extends HitBaseView<HitTriageTemplate, HitTriageApp> {
       async () => {
         await onTemmplateChange();
       });
-    const createNewtemplateButton = ui.icons.add(() => {
-      this.createNewTemplate();
+    const createNewtemplateButton = ui.icons.add(async () => {
+      ui.setUpdateIndicator(this.root, true);
+      await this.createNewTemplate();
+      ui.setUpdateIndicator(this.root, false);
     }, i18n.createNewTemplate);
     templatesInput.addOptions(createNewtemplateButton);
 
@@ -153,8 +166,10 @@ export class InfoView extends HitBaseView<HitTriageTemplate, HitTriageApp> {
     this.app.campaign = campaign;
   }
 
-  private async getNewCampaignAccordeon(template: HitTriageTemplate) {
-    const {root, promise, cancelPromise} = newCampaignAccordeon(template);
+  private async getNewCampaignAccordeon(template: HitTriageTemplate, campaignDetailsDiv: HTMLElement) {
+    ui.setUpdateIndicator(campaignDetailsDiv, true);
+    const {root, promise, cancelPromise} = await newCampaignAccordeon(template, this.dataSourceFunctionsMap);
+    ui.setUpdateIndicator(campaignDetailsDiv, false);
     promise.then(async (camp) => {
       this.app.dataFrame = camp.df;
       this.app._fileInputType = camp.type;
@@ -170,7 +185,7 @@ export class InfoView extends HitBaseView<HitTriageTemplate, HitTriageApp> {
   }
 
   private async createNewTemplate() {
-    const newTemplateAccordeon = await createTemplateAccordeon();
+    const newTemplateAccordeon = await createTemplateAccordeon(this.dataSourceFunctionsMap);
     // hideComponents(toRemove);
     // $(containerDiv).empty();
     // $(templateInputDiv).empty();