+++ packages/UsageAnalysis/src/test-track/app.ts
@@ -2,34 +2,25 @@ import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 
-import {Octokit} from 'octokit';
-import {getIcon} from './utils';
-import {decode} from 'js-base64';
+import {getIcon, getStatusIcon, Status} from './utils';
+import {_package} from '../package';
 
-const octokit = new Octokit();
-
-interface TestCase {
+interface TestCase extends Options {
   name: string;
+  path: string;
   text: string;
+  status: Status | null;
+  icon: HTMLElement;
+  reason: HTMLElement;
 }
 
-interface Category {
+interface Category extends Options {
   name: string;
   children: (TestCase | Category)[];
 }
 
-interface OctokitResponse {
-  // size: number;
-  name: string;
-  path: string;
-  type: 'dir' | 'file' | 'submodule' | 'symlink';
-  content?: string;
-  // sha: string;
-  // url: string;
-  // git_url: string | null;
-  // html_url: string | null;
-  // download_url: string | null;
-  // _links: {};
+interface Options {
+  order?: number;
 }
 
 export class TestTrack extends DG.ViewBase {
@@ -38,7 +29,10 @@ export class TestTrack extends DG.ViewBase {
   inited: boolean = false;
   testCaseDiv: HTMLDivElement;
   currentNode: DG.TreeViewNode | DG.TreeViewGroup;
-  objTree: Category[] = [];
+  map: {[key: string]: (Category | TestCase)} = {};
+  list: Category[] = [];
+  expanded: boolean = false;
+  version: string = grok.shell.build.client.version;
 
   public static getInstance(): TestTrack {
     if (!TestTrack.instance)
@@ -64,24 +58,46 @@ export class TestTrack extends DG.ViewBase {
     }
 
     // Generate tree
-    const response = await octokit.rest.repos.getContent({
-      owner: 'datagrok-ai',
-      repo: 'public',
-      path: 'packages/UsageAnalysis/files',
-    });
-    const dirs = (response.data as OctokitResponse[]).filter((f) => f.type === 'dir');
-    for (const dir of dirs)
-      this.objTree.push(await this.processDirRecursive(dir));
-    this.objTree.forEach((obj) => this.initTreeGroupRecursive(obj, this.tree));
+    const filesP = _package.files.list('Test Track', true);
+    const history: DG.DataFrame = await grok.functions.call('UsageAnalysis:TestTrack', {version: this.version});
+    for (const row of history.rows) {
+      const path = row.get('path');
+      const status: Status = row.get('status');
+      const reason: string = row.get('reason');
+      const el: TestCase = {name: '', path, text: '', status,
+        icon: status ? ui.div(getStatusIcon(status)) : ui.div(), reason: ui.div(reason, 'tt-reason')};
+      this.map[path] = el;
+    }
+    const files = await filesP;
+    const p: Promise<void>[] = [];
+    for (const file of files) {
+      if (file.isDirectory)
+        this.processDir(file);
+      else
+        p.push(this.processFile(file));
+    }
+    await Promise.all(p);
+    this.list.forEach((c) => this.sortCategoryRecursive(c));
+    this.list.forEach((obj) => this.initTreeGroupRecursive(obj, this.tree));
 
     // Ribbon
-    const plus = ui.button(getIcon('plus', {style: 'fas'}), () => {
-    }, 'Add test case');
-    plus.classList.add('tt-ribbon-button');
+    const gh = ui.button(getIcon('github', {style: 'fab'}), () => {
+      window.open('https://github.com/datagrok-ai/public/tree/master/packages/UsageAnalysis/files/Test Track',
+        '_blank')?.focus();
+    }, 'Edit list');
+    gh.classList.add('tt-ribbon-button');
     const report = ui.button(getIcon('tasks', {style: 'fas'}), () => {
     }, 'Generate report');
     report.classList.add('tt-ribbon-button');
-    const ribbon = ui.divH([plus, report]);
+    const ec = ui.button(getIcon('sort', {style: 'fas'}), () => {
+      this.expanded = !this.expanded;
+      this.tree.items.forEach((n: DG.TreeViewGroup | DG.TreeViewNode) => {
+        if (n.constructor === DG.TreeViewGroup)
+          n.expanded = this.expanded;
+      });
+    }, 'Expand/collapse');
+    report.classList.add('tt-ribbon-button');
+    const ribbon = ui.divH([gh, report, ec]);
 
     // Test case div
     this.tree.onSelectedNodeChanged.subscribe((node) => {
@@ -98,44 +114,129 @@ export class TestTrack extends DG.ViewBase {
     this.inited = true;
   }
 
-  async processDirRecursive(dir: OctokitResponse): Promise<Category> {
-    const c: Category = {name: dir.name, children: []};
-    const data = (await octokit.rest.repos.getContent({
-      owner: 'datagrok-ai',
-      repo: 'public',
-      path: dir.path,
-    })).data as OctokitResponse[];
-    for (const child of data) {
-      let res: Category | TestCase;
-      if (child.type === 'dir')
-        res = await this.processDirRecursive(child);
-      else if (child.type === 'file')
-        res = await this.processFile(child);
-      else continue;
-      c.children.push(res);
-    }
-    return c;
+  processDir(dir: DG.FileInfo): void {
+    const el: Category = {name: dir.name, children: []};
+    const pathL = dir.path.split('/').slice(2);
+    if (pathL.length > 1) {
+      const parent = this.map[pathL.slice(0, -1).join(': ')] as Category;
+      parent.children.push(el);
+    } else
+      this.list.push(el);
+    this.map[pathL.join(': ')] = el;
   }
 
-  async processFile(file: OctokitResponse): Promise<TestCase> {
-    const tc: TestCase = {name: file.name.replace(/\.[^/.]+$/, ''), text: ''};
-    const data = (await octokit.rest.repos.getContent({
-      owner: 'datagrok-ai',
-      repo: 'public',
-      path: file.path,
-    })).data as OctokitResponse;
-    if (data.content)
-      tc.text = decode(data.content);
-    return tc;
+  async processFile(file: DG.FileInfo): Promise<void> {
+    const pathL = file.path.replace(/\.[^.]+$/, '').split('/').slice(2);
+    if (pathL.length < 2)
+      grok.shell.error('Root test case');
+    const parent = this.map[pathL.slice(0, -1).join(': ')] as Category;
+    const [text, jsonS] = (await _package.files.readAsText(file)).split('\r\n---\r\n', 2);
+    const path = pathL.join(': ');
+    const elOld: TestCase | undefined = this.map[path] as TestCase;
+    const status = elOld ? elOld.status : null;
+    const reason = elOld ? elOld.reason : ui.div('', 'tt-reason');
+    const icon = elOld ? elOld.icon : ui.div();
+    let el: TestCase = {name: file.name.replace(/\.[^.]+$/, ''), path, text, status, reason, icon};
+    if (jsonS) {
+      const json: Options = JSON.parse(jsonS);
+      el = {...json, ...el};
+    }
+    parent.children.push(el);
+    this.map[path] = el;
+  }
+
+  sortCategoryRecursive(cat: Category): void {
+    if (!cat.children.length) return;
+    const cats = cat.children.filter((c) => 'children' in c) as Category[];
+    cat.children.sort((a, b) => {
+      if ('children' in a && !('children' in b)) return -1;
+      if ('children' in b && !('children' in a)) return 1;
+      if (a.order === undefined && b.order === undefined)
+        return a.name.localeCompare(b.name);
+      if (a.order === undefined) return 1;
+      if (b.order === undefined) return -1;
+      return a.order > b.order ? 1 : -1;
+    });
+    cats.forEach((c) => this.sortCategoryRecursive(c));
   }
 
   initTreeGroupRecursive(obj: Category | TestCase, parent: DG.TreeViewGroup): void {
     if ('text' in obj) {
-      parent.item(obj.name, obj);
+      const node = parent.item(obj.name, obj);
+      this.setContextMenu(node);
+      (node.value.reason as HTMLElement).ondblclick = () => this.showChangeReasonDialog(node, node.value.status);
+      node.captionLabel.after(node.value.reason);
+      node.captionLabel.after(node.value.icon);
       return;
     }
     const group = parent.getOrCreateGroup(obj.name, obj, false);
     for (const child of obj.children)
       this.initTreeGroupRecursive(child, group);
+  }
+
+  setContextMenu(node: DG.TreeViewNode): void {
+    node.captionLabel.addEventListener('contextmenu', (e) => {
+      DG.Menu.popup()
+        .group('Status').items(['Passed', 'Failed', 'Skipped'],
+          (i) => {
+            const status = i.toLowerCase() as Status;
+            if (node.value.status === status) return;
+            if (status === 'failed' || status === 'skipped')
+              this.showChangeNodeStatusDialog(node, status);
+            else
+              this.changeNodeStatus(node, status);
+          },
+          {radioGroup: 'Status', isChecked: (i) => i.toLowerCase() === (node.value.status ?? 'empty')})
+        .endGroup()
+        .item('Edit', async () => {
+          window.open(`https://github.com/datagrok-ai/public/edit/master/packages/UsageAnalysis/files/Test Track/${
+            node.value.path.replaceAll(': ', '/')}.md`, '_blank')?.focus();
+        })
+        .show();
+      e.preventDefault();
+      e.stopPropagation();
+    });
+  }
+
+  changeNodeStatus(node: DG.TreeViewNode, status: Status, reason?: string): void {
+    node.value.status = status;
+    node.value.icon.innerHTML = '';
+    node.value.reason.innerText = '';
+    const icon = getStatusIcon(status);
+    node.value.icon.append(icon);
+    if (status === 'failed' || status === 'skipped')
+      node.value.reason.innerText = reason;
+    const params = {success: status === 'passed', result: reason ?? '', skipped: status === 'skipped', type: 'manual',
+      category: node.value.path.replace(/:\s[^:]+$/, ''), test: node.text, version: this.version};
+    grok.log.usage(node.value.path, params, `test-manual ${node.value.path}`);
+  }
+
+  showChangeNodeStatusDialog(node: DG.TreeViewNode, status: 'failed' | 'skipped'): void {
+    const dialog = ui.dialog(status === 'failed' ? 'Specify ticket' : 'Specify skip reason');
+    const input = ui.textInput(status === 'failed' ? 'Key' : 'Reason', '', () => {});
+    input.nullable = false;
+    dialog.add(input);
+    dialog.onOK(() => this.changeNodeStatus(node, status, input.value));
+    dialog.show({resizable: true});
+  }
+
+  changeNodeReason(node: DG.TreeViewNode, reason: string, status: Status): void {
+    if (status !== node.value.status) {
+      grok.shell.warning('Test case status was changed');
+      return;
+    }
+    node.value.reason.innerText = reason;
+    const params = {success: status === 'passed', result: reason, skipped: status === 'skipped', type: 'manual',
+      category: node.value.path.replace(/:\s[^:]+$/, ''), test: node.text, version: this.version};
+    grok.log.usage(node.value.path, params, `test-manual ${node.value.path}`);
+  }
+
+  showChangeReasonDialog(node: DG.TreeViewNode, status: Status): void {
+    const dialog = ui.dialog(status === 'failed' ? 'Edit ticket' : 'Edit skip reason');
+    const input = ui.textInput(status === 'failed' ? 'Key' : 'Reason', node.value.reason.innerText, () => {});
+    input.nullable = false;
+    dialog.add(input);
+    dialog.onOK(() => this.changeNodeReason(node, input.value, status));
+    dialog.show({resizable: true});
   }
 };