+++ tools/bin/commands/test.ts
@@ -14,7 +14,7 @@ import * as testUtils from '../utils/test-utils';
 
 export function test(args: TestArgs): boolean {
   const options = Object.keys(args).slice(1);
-  const commandOptions = ['host', 'csv', 'gui', 'catchUnhandled',
+  const commandOptions = ['host', 'csv', 'gui', 'catchUnhandled', 'platform',
     'report', 'skip-build', 'skip-publish', 'category', 'record', 'verbose', 'benchmark'];
   const nArgs = args['_'].length;
   const curDir = process.cwd();
@@ -56,6 +56,11 @@ export function test(args: TestArgs): boolean {
     console.log('Environment variable `TARGET_PACKAGE` is set to', process.env.TARGET_PACKAGE);
   } else {
     color.error('Invalid package name. Set the `name` field in `package.json`');
+    return false;
+  }
+
+  if (args.platform && process.env.TARGET_PACKAGE !== 'ApiTests') {
+    color.error('--platform flag can only be used in the ApiTests package');
     return false;
   }
 
@@ -121,7 +126,7 @@ export function test(args: TestArgs): boolean {
     }
 
     function runTest(timeout: number, options: {category?: string, catchUnhandled?: boolean,
-      report?: boolean, record?: boolean, verbose?: boolean, benchmark?: boolean} = {}): Promise<resultObject> {
+      report?: boolean, record?: boolean, verbose?: boolean, benchmark?: boolean, platform?: boolean} = {}): Promise<resultObject> {
       return testUtils.runWithTimeout(timeout, async () => {
         let consoleLog: string = '';
         if (options.record) {
@@ -141,7 +146,7 @@ export function test(args: TestArgs): boolean {
           if (options.benchmark)
             (<any>window).DG.Test.isInBenchmark = true;
           return new Promise<resultObject>((resolve, reject) => {        
-            (<any>window).grok.functions.call(`${targetPackage}:test`, {
+            (<any>window).grok.functions.call(`${targetPackage}:${options.platform ? 'testPlatform' : 'test'}`, {
               'category': options.category,
               'testContext': testContext,
             }).then((df: any) => {
@@ -206,7 +211,7 @@ export function test(args: TestArgs): boolean {
         throw e;
       }
 
-      const r = await runTest(7200000, {category: args.category, verbose: args.verbose,
+      const r = await runTest(7200000, {category: args.category, verbose: args.verbose, platform: args.platform,
         catchUnhandled: args.catchUnhandled, report: args.report, record: args.record, benchmark: args.benchmark});
 
       if (r.csv && args.csv) {
@@ -254,5 +259,6 @@ interface TestArgs {
   'skip-build'?: boolean,
   'skip-publish'?: boolean,
   verbose?: boolean,
-  benchmark?: boolean
+  benchmark?: boolean,
+  platform?: boolean
 }